+++ CHANGELOG.md
@@ -18,6 +18,8 @@
 + Added **Odonatas Highlighter**
   + Highlight the small Odonatas flying around the trees while holding a Empty Odonata Bottle in the hand.
   + Customize the color
++ Added **Agaricus Cap** countdown
+  + Counts down the time until Agaricus Cap (Mushroom) changes color from brown to red and is breakable.
 
 ### Changes
 + Option to show the yaw and pitch display outside garden or while not holding a farming tool (Contributed by CalMWolfs)